api-gateway/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
requests==2.28.2
django-cors-headers==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
dj-database-url==1.2.0